+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 0.5.0
 
 - Added a ``with`` tag. A tag that extends the local namespace for the duration of its
   block.
+- Removed depreciated ``env`` argument to the ``JSON`` and ``Translate`` filters.
 
 Version 0.4.0
 -------------